2009-10-26 15:29 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/contrib/hbqt/qtcore/QLocale.cpp   * harbour/contrib/hbqt/qtcore/QByteArray.cpp     ! fixed casting for 64bit builds - in all 64bit builds except Win64       LONG is 64bit integer and HB_LONG is declared as LONG not LONGLONG
  * harbour/contrib/hbwin/win_dll.c
    % minor optimization
    * removed unnecessary casting

+++ harbour/contrib/hbqt/qtcore/QByteArray.cpp
@@ -684,7 +684,7 @@ HB_FUNC( QT_QBYTEARRAY_SETNUM_3 )
  */
 HB_FUNC( QT_QBYTEARRAY_SETNUM_4 )
 {
-   hb_retptrGC( hbqt_ptrTOgcpointer( new QByteArray( hbqt_par_QByteArray( 1 )->setNum( hb_parnint( 2 ), ( HB_ISNUM( 3 ) ? hb_parni( 3 ) : 10 ) ) ), release_QByteArray ) );
+   hb_retptrGC( hbqt_ptrTOgcpointer( new QByteArray( hbqt_par_QByteArray( 1 )->setNum( ( qulonglong ) hb_parnint( 2 ), ( HB_ISNUM( 3 ) ? hb_parni( 3 ) : 10 ) ) ), release_QByteArray ) );
 }
 
 /*
@@ -692,7 +692,7 @@ HB_FUNC( QT_QBYTEARRAY_SETNUM_4 )
  */
 HB_FUNC( QT_QBYTEARRAY_SETNUM_5 )
 {
-   hb_retptrGC( hbqt_ptrTOgcpointer( new QByteArray( hbqt_par_QByteArray( 1 )->setNum( hb_parnint( 2 ), ( HB_ISNUM( 3 ) ? hb_parni( 3 ) : 10 ) ) ), release_QByteArray ) );
+   hb_retptrGC( hbqt_ptrTOgcpointer( new QByteArray( hbqt_par_QByteArray( 1 )->setNum( ( qulonglong ) hb_parnint( 2 ), ( HB_ISNUM( 3 ) ? hb_parni( 3 ) : 10 ) ) ), release_QByteArray ) );
 }
 
 /*
